requirements.txt
pandas>=1.5.0
requests>=2.28.0
pathlib>=1.0.1
openai>=1.0.0
anthropic>=0.34.0
pyyaml>=6.0.1

